enhancement(Viewer): Style improvements, components refactoring
