+++ src/containers/ReviewablesList/ReviewablesUpload.tsx
@@ -3,6 +3,7 @@ import api from '@api'
 import { toast } from 'react-toastify'
 import { FC, useState, DragEvent, ChangeEvent, useRef, useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
+import clsx from 'clsx'
 
 import { Icon } from '@ynput/ayon-react-components'
 import { $Any } from '@types'
@@ -13,46 +14,7 @@ import { toggleUpload } from '@state/viewer'
 
 // components
 import ReviewableProgressCard, { ReviewableProgress } from '@components/ReviewableProgressCard'
-import * as Styled from './ReviewablesList.styled'
-import styled from 'styled-components'
-
-export const ReviewablesList = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: var(--base-gap-small);
-  min-height: 100%;
-  padding-bottom: var(--padding-m);
-  width: 100%;
-  height: 100%;
-`
-
-export const Upload = styled.div`
-  position: relative;
-  min-height: 50px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  user-select: none;
-
-  background-color: var(--md-sys-color-surface-container-low);
-
-  border-radius: var(--border-radius-xxl);
-  border: 2px dashed var(--md-sys-color-outline-variant);
-  color: var(--md-sys-color-outline-variant);
-
-  &:hover {
-    background-color: var(--md-sys-color-surface-container-low-hover);
-    border-color: var(--md-sys-color-outline);
-    color: var(--md-sys-color-outline);
-  }
-
-  input {
-    position: absolute;
-    inset: 0;
-    opacity: 0;
-    cursor: pointer;
-  }
-`
+import * as Styled from './ReviewablesUpload.styled'
 
 interface ReviewableUploadProps {
   projectName: string
@@ -232,7 +194,10 @@ const ReviewableUpload: FC<ReviewableUploadProps> = ({
   return (
     <>
       {!isDraggingFile && (
-        <ReviewablesList onDragEnter={() => setIsDraggingFile(true)}>
+        <Styled.ReviewablesList
+          onDragEnter={() => setIsDraggingFile(true)}
+          className={clsx(variant)}
+        >
           <>
             {children}
             {/* uploading items */}
@@ -246,12 +211,12 @@ const ReviewableUpload: FC<ReviewableUploadProps> = ({
             ))}
 
             {/* upload button */}
-            <Upload className="upload" style={{ height: variant == 'large' ? '80px' : ''}}>
+            <Styled.Upload id="upload" className={clsx('upload', variant)} style={{}}>
               <span>Drop or click to upload</span>
               <input type="file" multiple onChange={handleInputChange} ref={inputRef} />
-            </Upload>
+            </Styled.Upload>
           </>
-        </ReviewablesList>
+        </Styled.ReviewablesList>
       )}
 
       {isDraggingFile && (
